cmd/etc_sample/etc_sample_test.go
package etc_sample

import (
    "os"
    "sync"
    "testing"
    "time"
)

// TestReadFile tests the ReadFile function
func TestReadFile(t *testing.T) {
    // Create a temporary test file
    tempFile, err := os.CreateTemp("", "test")
    if err != nil {
        t.Fatalf("Failed to create temp file: %s", err)
    }
    defer os.Remove(tempFile.Name())

    _, err = ReadFile(tempFile.Name())
    if err != nil {
        t.Errorf("Failed to read file: %s", err)
    }
}

// TestGoroutineChannelUsage tests the GoroutineChannelUsage function
func TestGoroutineChannelUsage(t *testing.T) {
    // This function is simple and may not require a detailed test
    // You can test if it runs without deadlocks or panics
}

// TestPersonSpeak tests the Speak method of the Person struct
func TestPersonSpeak(t *testing.T) {
    bob := Person{}
    expected := "Hello"
    actual := bob.Speak()
    if actual != expected {
        t.Errorf("Expected %s, got %s", expected, actual)
    }
}

// TestProcessFile tests the ProcessFile function
func TestProcessFile(t *testing.T) {
    // Create a temporary test file
    tempFile, err := os.CreateTemp("", "test")
    if err != nil {
        t.Fatalf("Failed to create temp file: %s", err)
    }
    defer os.Remove(tempFile.Name())

    // Call ProcessFile with the temp file
    ProcessFile(tempFile.Name())
    // You might want to check if the file was processed correctly
}

// TestIncrement tests the increment function
func TestIncrement(t *testing.T) {
    var mu sync.Mutex
    counter := 0

    for i := 0; i < 1000; i++ {
        go func() {
            mu.Lock()
            counter++
            mu.Unlock()
        }()
    }
    time.Sleep(time.Second) // Wait for goroutines to finish

    if counter != 1000 {
        t.Errorf("Expected counter to be 1000, got %d", counter)
    }
}

// Additional tests can be added for other functions or specific scenarios
